[refactor] Mission3: Step2: Person -> Customer, Barista 상속화
- Person을 통해, Customer와 Barista가 상속을 받아 클래스를 구현할수 있도록 반영했습니다.
